+++ packages/radix-ui-themes/src/components/skeleton.props.ts
@@ -1,7 +1,10 @@
+import { widthPropDefs, heightPropDefs } from '../helpers';
 import type { PropDef } from '../helpers';
 
 const skeletonPropDefs = {
   loading: { type: 'boolean', default: true },
+  ...widthPropDefs,
+  ...heightPropDefs,
 } satisfies {
   loading: PropDef<boolean>;
 };