+++ packages/radix-ui-themes/src/components/skeleton.tsx
@@ -2,7 +2,15 @@ import * as React from 'react';
 import classNames from 'classnames';
 import { Slot } from '@radix-ui/react-slot';
 import { skeletonPropDefs } from './skeleton.props';
-import { extractMarginProps, withMarginProps } from '../helpers';
+import {
+  extractHeightProps,
+  extractMarginProps,
+  extractWidthProps,
+  getHeightStyles,
+  getWidthStyles,
+  mergeStyles,
+  withMarginProps,
+} from '../helpers';
 
 import type { MarginProps, GetPropDefTypes } from '../helpers';
 
@@ -14,12 +22,18 @@ interface SkeletonProps
     SkeletonOwnProps {}
 const Skeleton = React.forwardRef<SkeletonElement, SkeletonProps>((props, forwardedRef) => {
   const { rest: marginRest, ...marginProps } = extractMarginProps(props);
+  const { rest: widthRest, ...widthProps } = extractWidthProps(marginRest);
+  const { rest: heightRest, ...heightProps } = extractHeightProps(widthRest);
   const {
     className,
     children,
+    style,
     loading = skeletonPropDefs.loading.default,
     ...skeletonProps
-  } = marginRest;
+  } = heightRest;
+
+  const [widthClassNames, widthCustomProperties] = getWidthStyles(widthProps);
+  const [heightClassNames, heightCustomProperties] = getHeightStyles(heightProps);
 
   if (!loading) return <>{children}</>;
 
@@ -29,9 +43,16 @@ const Skeleton = React.forwardRef<SkeletonElement, SkeletonProps>((props, forwar
     <Tag
       ref={forwardedRef}
       aria-hidden
-      className={classNames('rt-Skeleton', className, withMarginProps(marginProps))}
+      className={classNames(
+        'rt-Skeleton',
+        className,
+        widthClassNames,
+        heightClassNames,
+        withMarginProps(marginProps)
+      )}
       data-inline-skeleton={React.isValidElement(children) ? undefined : true}
       tabIndex={-1}
+      style={mergeStyles(widthCustomProperties, heightCustomProperties, style)}
       // Workaround to use `inert` until https://github.com/facebook/react/pull/24730 is merged.
       {...{ inert: true ? '' : undefined }}
       {...skeletonProps}